requirements.txt
streamlit==1.32.0
openai==0.28.1
speechrecognition
pydub==0.25.1
numpy==1.26.4
pandas==2.2.1
whisper==1.1.10
tiktoken==0.6.0
python-dotenv==1.0.1
pyannote.audio==3.1.1
torch==2.2.1
st_audiorecorder
google-generativeai 